Move haptics to a separate utility class
